Removing totally useless file: Arggh Excel! Arghh!
From Dan: It seems that $ospree was some strange, alien file. When I tried to sync this morning, git reported a conflict with it. Then I tried to open it, it failed to open and then disappeared, seemingly forever. It was only 173 bytes, so I assume whatever it was, it was not very important.