Removing revisiting folder
This folder was created by DB Flynn to store files that mark which studies in OPSPREE we needed to revisit and check (hence the name revisiting). As of 26 Jan 2017 the group went through this folder and agreed that either (a) all these changes had been dealt with or (2) we have new code and new files that flag which data in OSPREE to double check.